I didn't really have any commits because cherry-pick actually did it for me.  But I wanted to record in the history the command I ran.
git cherry-pick -x 16493ebc03ffc665cac812fbfdce7d35864bb384

What's crazy is that it seems like the graph magically restructured itself.  That commit was between two other commits on master.  Now it's behind the one that was behind it.  I don't know if that's a graph rendering issue, or if cherry-pick actually did that.

While I was at it, I added some lorem.  I love that lorem.